Enable CI for macOS runners
Unfortunately the store redirection used on Linux to enable caching of
the nix store is not supported on macOS as you cannot easily `chroot` in
userspace. While we may be able to do better in the future, for now we
just do not cache the nix store on macOS, and install all of the
development dependencies from scratch.