Add support for our temporary target (#68)
For the current phase of work we are requiring that input LLVM IR be
built to target the `aarch64-unknown-none-softfloat` target. This means
that it can make no assumptions about either the host operating system
(as there is none), nor the target ABI (meaning that we fall back to
platform codegen ABI convention, and hence that `rustc` will not do
anything funky here).

We add support in the nix derivation (`flake.nix`) so that this target
is automatically made available in the sandbox. It may be necessary to
add it at the system level as well, depending on how your development
environment is set up.