Implement the skeleton of a pass infrastructure
While it is incomplete, an initial skeleton for the pass management
infrastructure ensures that we do not paint ourselves into any corners
that might be hard to design our way out of later. It provides the
infrastructure for creating and managing passes, as well as shuffling
pass data between passes.

Please note that this we make use of self-referential structs via
Ouroboros to better encapsulate the LLVM context and the modules that
exist in it.